requirements.txt
pymongo[srv]>=4.6.1
voyageai==0.2.1
python-dotenv==1.0.0
anthropic==0.18.1
numpy>=1.24
pandas>=2.0
openai==1.12.0
cvxpy>=1.3.0
scipy>=1.10.0
ecos>=2.0.0
scikit-learn>=1.2.0
pyyaml>=6.0
pydantic>=2.0
pytest>=7.4
pytest-cov>=4.1
prometheus-client>=0.20

# Optional: Quantum Computing Support
# Uncomment to enable D-Wave quantum optimization
# dwave-ocean-sdk>=6.0.0
